Update docker-compose.yml
This is a temporary change made to make docker-compose functional before the NodeBB team adds the Automated Build to Docker Hub as nodebb/nodebb.  Again, please consider this only a placeholder until then.  It serves to build the Dockerfile locally, so that the user can get pure NodeBB satisfaction.  Once the change has been made, change line 2 from build: . to image: nodebb/nodebb

That's a wrap!